Service admin refactoring
